fix: not exiting alt buffer when audio playback ends
i am so stupid sometimes:
i thought this problem originated from input.rs where we'd not restore
the terminal, but thats not it. it seems to do that anyways (probably my
shell, fish? if not, i have a stashed change that would fix it)

and then i realized its because of one single `break` after staring at
my code, searching everywhere for like 30 minutes